Moved the PDOM into the Core. Started introduction of ILanguage, the mechanism of language extensibility for the CDT.

+++ core/org.eclipse.cdt.core/src/org/eclipse/cdt/core/dom/PDOM.java
@@ -10,15 +10,21 @@
  *******************************************************************************/
 package org.eclipse.cdt.core.dom;
 
+import org.eclipse.cdt.core.CCorePlugin;
 import org.eclipse.cdt.core.model.CoreModel;
 import org.eclipse.cdt.core.model.IElementChangedListener;
 import org.eclipse.cdt.internal.core.dom.NullPDOMProvider;
+import org.eclipse.cdt.internal.core.pdom.PDOMManager;
 import org.eclipse.core.resources.IProject;
+import org.eclipse.core.resources.ProjectScope;
 import org.eclipse.core.runtime.CoreException;
 import org.eclipse.core.runtime.IConfigurationElement;
 import org.eclipse.core.runtime.IExtension;
 import org.eclipse.core.runtime.IExtensionPoint;
 import org.eclipse.core.runtime.Platform;
+import org.eclipse.core.runtime.preferences.IEclipsePreferences;
+import org.eclipse.core.runtime.preferences.IScopeContext;
+import org.osgi.service.prefs.BackingStoreException;
 
 /**
  * @author Doug Schaefer
@@ -56,6 +62,37 @@ public class PDOM {
     	return pdomProvider;
     }
 
+    private static final String PDOM_NODE = CCorePlugin.PLUGIN_ID + ".pdom"; //$NON-NLS-1$ 
+    private static final String ENABLED_KEY = "enabled"; //$NON-NLS-1$
+    	
+    private static IEclipsePreferences getPreferences(IProject project) {
+    	IScopeContext projectScope = new ProjectScope(project);
+    	if (projectScope == null)
+    		return null;
+    	else
+    		return projectScope.getNode(PDOM_NODE);
+    }
+    
+    public static boolean isEnabled(IProject project) {
+    	IEclipsePreferences prefs = getPreferences(project);
+    	if (prefs == null)
+    		return false;
+    	else
+    		return prefs.getBoolean(ENABLED_KEY, false);
+    }
+    
+    public static void setEnabled(IProject project, boolean enabled) {
+    	IEclipsePreferences prefs = getPreferences(project);
+    	if (prefs == null)
+    		return;
+    	
+    	prefs.putBoolean(ENABLED_KEY, enabled);
+    	try {
+    		prefs.flush();
+    	} catch (BackingStoreException e) {
+    	}
+    }
+
     /**
      * Get the PDOM for the given project.
      * 
@@ -63,11 +100,15 @@ public class PDOM {
      * @return the PDOM for the project
      */
 	public static IPDOM getPDOM(IProject project) {
-		return getPDOMProvider().getPDOM(project);
+		if (isEnabled(project))
+			return PDOMManager.getInstance().getPDOM(project);
+		else
+			return null;
 	}
 
 	public static void deletePDOM(IProject project) throws CoreException {
-		getPDOMProvider().getPDOM(project).delete();
+		if (isEnabled(project))
+			getPDOMProvider().getPDOM(project).delete();
 	}
 	
 	/**
